k8s-cluster-ec2/requirements.txt
-e .
aws_cdk.core==1.120.0
aws_cdk.aws-iam==1.120.0
aws_cdk.aws-eks==1.120.0
aws-cdk.aws_ec2==1.120.0
pyyaml